Load static param and shape data under strokes collection.
