--- app/collections/data-sentences.js
@@ -1,19 +1,0 @@
-var GelatoCollection = require('gelato/collection');
-var DataSentence = require('models/data-sentence');
-
-/**
- * @class DataSentences
- * @extends {GelatoCollection}
- */
-module.exports = GelatoCollection.extend({
-    /**
-     * @method initialize
-     * @constructor
-     */
-    initialize: function() {},
-    /**
-     * @property model
-     * @type {DataSentence}
-     */
-    model: DataSentence
-});
