+++ app/models/stroke.js
@@ -1,4 +1,6 @@
 var GelatoModel = require('gelato/model');
+var PromptCharacter = require('collections/prompt-character');
+var PromptStroke = require('models/prompt-stroke');
 
 /**
  * @class Stroke
@@ -9,5 +11,43 @@ module.exports = GelatoModel.extend({
      * @property idAttribute
      * @type {String}
      */
-    idAttribute: 'rune'
+    idAttribute: 'rune',
+    /**
+     * @method getPromptCharacter
+     * @returns {PromptCharacter}
+     */
+    getPromptCharacter: function() {
+        var character = new PromptCharacter();
+        var variations = this.clone().get('strokes');
+        var rune = this.get('rune');
+        var targets = [];
+        for (var a = 0, lengthA = variations.length; a < lengthA; a++) {
+            var target = new PromptCharacter();
+            var targetVariation = variations[a];
+            var strokePosition = 0;
+            target.position = a;
+            for (var b = 0, lengthB = targetVariation.length; b < lengthB; b++) {
+                var stroke = new PromptStroke();
+                var strokeData = targetVariation[b];
+                var strokeId = strokeData[0];
+                var strokeParam = this.collection.params.findWhere({strokeId: strokeId});
+                var strokeContains = strokeParam.get('contains');
+                stroke.set({
+                    contains: strokeContains,
+                    data: strokeData,
+                    id: strokePosition + '-' + strokeId,
+                    position: strokePosition,
+                    shape: this.collection.shapes.get(strokeId),
+                    strokeId: strokeId,
+                    tone: rune === 'tones' ? a + 1 : null
+                });
+                strokePosition += strokeContains.length || 1;
+                target.add(stroke);
+            }
+            targets.push(target);
+        }
+        character.targets = targets;
+        character.writing = rune;
+        return character;
+    }
 });
